requirements.txt
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
google-generativeai>=0.8.0
web3>=6.11.3
eth-account>=0.10.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.14
aiosqlite>=0.19.0
greenlet>=3.0.0
